refactor: merge internal/vault into pkg/vault
Same Phase 0.1 pattern as #326 / #335 / #336. The Vault interface
already lived in pkg/vault; the implementations (LocalVault, GCPVault)
lived in internal/vault. External consumers (clawvisor-cloud, where
the runtime proxy fleet wires a master-keyed LocalVault as the
fallback for OrgAwareVault) couldn't construct LocalVault directly.

This commit merges the internal files into the existing pkg/vault
package. Both directories were already package vault, so the merge
drops the pkgvault self-import and converts pkgvault.ErrNotFound to
ErrNotFound (now defined in the same package). gofmt cleans up the
resulting import blocks.

Caller imports updated across 8 files (test files in
internal/api/handlers/, internal/api/, internal/notify/telegram/,
internal/e2e/harness/, plus pkg/clawvisor/defaults.go and two
pkg/runtime/proxy/ test files).

No API or behavior change. LocalVault, NewLocalVault,
NewLocalVaultFromKey, NewLocalVaultFromKeyWithDB, GCPVault,
NewGCPVault, ResolveKey all keep their signatures.

Verified: go vet, go build, go test ./... all pass.
