+++ pkg/vault/local.go
@@ -13,8 +13,6 @@ import (
 	"os"
 
 	"github.com/google/uuid"
-
-	pkgvault "github.com/clawvisor/clawvisor/pkg/vault"
 )
 
 // LocalVault encrypts credentials with AES-256-GCM and stores them in the
@@ -94,7 +92,7 @@ func (v *LocalVault) Get(ctx context.Context, userID, serviceID string) ([]byte,
 	var encrypted, iv, authTag string
 	err := v.db.QueryRowContext(ctx, q, userID, serviceID).Scan(&encrypted, &iv, &authTag)
 	if errors.Is(err, sql.ErrNoRows) {
-		return nil, pkgvault.ErrNotFound
+		return nil, ErrNotFound
 	}
 	if err != nil {
 		return nil, fmt.Errorf("vault get: %w", err)
